requirements.txt
# for compiling
requests
python-dotenv
tinify